creating user endpoint and seperating from the auth
