+++ src/routes/AuthRouter.js
@@ -1,13 +1,6 @@
 const authRouter = require("express").Router();
-const AuthController = require("../controllers/UserController");
-const authMid = require("../middlewares/AuthMiddleware");
+const AuthController = require("../controllers/AuthController");
 
 module.exports = authRouter
   .post("/register", AuthController.register)
-  .post("/login", AuthController.login)
-  .delete(
-    "/deleteUser/:userId",
-    authMid.authorizeLogin,
-    authMid.authorizeRole([1]),
-    AuthController.deleteUserById
-  );
+  .post("/login", AuthController.login);
