+++ src/routes/UserRouter.js
@@ -1,0 +1,10 @@
+const userRouter = require("express").Router();
+const UserController = require("../controllers/UserController");
+const authMid = require("../middlewares/AuthMiddleware");
+
+module.exports = userRouter.delete(
+  "/deleteUser/:userId",
+  authMid.authorizeLogin,
+  authMid.authorizeRole([1]),
+  UserController.deleteUserById
+);
