+++ src/routes/index.js
@@ -1,9 +1,11 @@
 const authRouter = require("./AuthRouter");
+const userRouter = require("./AuthRouter");
 const courseRouter = require("./CourseRouter");
 const paymentRouter = require("./PaymentRouter");
 
 module.exports = {
   authRouter,
+  userRouter,
   courseRouter,
   paymentRouter,
 };
